PrefLib 240 \(\times\) 4 Data

Data of 240 cities across the globe ranked on four criteria from the ED-00015-001.soc dataset in the PrefLib repository. The first column contains the object names and each subsequent column is a complete ranking of the 240 objects with no ties) .

data(data240x4)

Format

A data frame with 240 rows and 5 columns:

Object

object name

Ranking 1

ranking on the first criterion

Ranking 2

ranking on the second criterion

Ranking 3

ranking on the third criterion

Ranking 4

ranking on the fourth criterion

Source

https://www.preflib.org/

References

Badal, P. S., & Das, A. (2018). Efficient algorithms using subiterative convergence for Kemeny ranking problem. Computers & Operations Research, 98, 198-210. doi:10.1016/j.cor.2018.06.007

Mattei, N., & Walsh, T. (2013, November). Preflib: A library for preferences https://www.preflib.org/. In International conference on algorithmic decision theory (pp. 259-270). Springer, Berlin, Heidelberg.

Examples

data(data240x4)
input_rkgs <- t(as.matrix(data240x4[, -1]))
obj_names <- data240x4[,1]

# Determine the mean seed ranking
mean_seed(input_rkgs)
#>   [1]  39  35  49   1  48  34   5   2   7  10   4  37   9   3  58  33   8  59
#>  [19]  14  45  12  60  65  15  17  11  23  16  22  18  66  38   6  24  50  77
#>  [37]  27  83  29  28  13  21  26  81  78  84  67  76  30  41  87  36  20 145
#>  [55]  82  31  53  54  92  19  68  95  44  25  73  93  90  64  42  51  55  47
#>  [73]  56 113  46  75  43  61  52  72 111  63  79 102 105 110  88  69  40 114
#>  [91] 125 103 115  32  62  57  98  80  86 117  71  74 109  89 128 162 119 130
#> [109]  91  94 138 120 137  85  96 134  97 100 133 101 135  99 183 154 176 107
#> [127] 106 104  70 147 118 151 108 131 122 148 121 116 169 188 165 153 126 159
#> [145] 123 177 167 127 141 157 149 124 129 158 146 164 161 142 143 132 171 136
#> [163] 144 139 152 197 189 170 190 175 184 160 112 180 155 150 156 191 140 168
#> [181] 210 179 166 163 187 198 193 195 174 182 201 186 217 196 222 220 206 204
#> [199] 219 185 181 192 221 172 202 227 233 178 199 173 200 203 194 223 211 208
#> [217] 216 209 207 239 213 214 218 215 224 238 226 228 225 240 229 232 230 212
#> [235] 234 231 235 236 237 205