Simulated 400 \(\times\) 15 Data

Data of 400 objects and 15 attributes in which the first column contains the object names and each subsequent column is a complete ranking of the 400 objects. This data set is generated from the 100 \(\times\) 15 dataset (see data50x15) by adding 100 to the ranks of the objects numbered 1 through 100 to get the ranks of objects numbered 101 through 200. Similarly, by adding 200 to obtain ranking 201 through 300, and by adding 300 to obtain ranking 301 through 400.

data(data400x15)

Format

A data frame with 400 rows and 16 columns:

Objects

object name

Ranking 1

ranking on the first attribute

Ranking 2

ranking on the second attribute

Ranking 3

ranking on the third attribute

Ranking 4

ranking on the fourth attribute

Ranking 5

ranking on the fifth attribute

Ranking 6

ranking on the sixth attribute

Ranking 7

ranking on the seventh attribute

Ranking 8

ranking on the eigth attribute

Ranking 9

ranking on the ninth attribute

Ranking 10

ranking on the tenth attribute

Ranking 11

ranking on the eleventh attribute

Ranking 12

ranking on the twelfth attribute

Ranking 13

ranking on the thirteenth attribute

Ranking 14

ranking on the fourteenth attribute

Ranking 15

ranking on the fifteenth attribute

Source

Badal, P. S., & Das, A. (2018). Efficient algorithms using subiterative convergence for Kemeny ranking problem. Computers & Operations Research, 98, 198-210. doi:10.1016/j.cor.2018.06.007

Examples

data(data400x15)
input_rkgs <- t(as.matrix(data400x15[, -1]))
obj_names <- data400x15[,1]

# Determine the mean seed ranking
mean_seed(input_rkgs)
#>   [1]  59  54  84  55  88  10  16  44  45   8  73  63  94  61  33  14  27  51
#>  [19]  41  28  23  38 100  87  43  58  18  97  25  47  71  78  39  96  77  62
#>  [37]  15  32   9  65  98  26  83   3  12  48  91  31  52  64  90  82  34  80
#>  [55]  66  36  37   4  30  89  35  95  20  24  40   2  67  21  93  42  76  50
#>  [73]  13  46  92   7  17  56  49  85  22  86  72  11  99  68   5  81  75  57
#>  [91]  60  53  29  74  19  70  79  69   6   1 159 154 184 155 188 110 116 144
#> [109] 145 108 173 163 194 161 133 114 127 151 141 128 123 138 200 187 143 158
#> [127] 118 197 125 147 171 178 139 196 177 162 115 132 109 165 198 126 183 103
#> [145] 112 148 191 131 152 164 190 182 134 180 166 136 137 104 130 189 135 195
#> [163] 120 124 140 102 167 121 193 142 176 150 113 146 192 107 117 156 149 185
#> [181] 122 186 172 111 199 168 105 181 175 157 160 153 129 174 119 170 179 169
#> [199] 106 101 259 254 284 255 288 210 216 244 245 208 273 263 294 261 233 214
#> [217] 227 251 241 228 223 238 300 287 243 258 218 297 225 247 271 278 239 296
#> [235] 277 262 215 232 209 265 298 226 283 203 212 248 291 231 252 264 290 282
#> [253] 234 280 266 236 237 204 230 289 235 295 220 224 240 202 267 221 293 242
#> [271] 276 250 213 246 292 207 217 256 249 285 222 286 272 211 299 268 205 281
#> [289] 275 257 260 253 229 274 219 270 279 269 206 201 359 354 384 355 388 310
#> [307] 316 344 345 308 373 363 394 361 333 314 327 351 341 328 323 338 400 387
#> [325] 343 358 318 397 325 347 371 378 339 396 377 362 315 332 309 365 398 326
#> [343] 383 303 312 348 391 331 352 364 390 382 334 380 366 336 337 304 330 389
#> [361] 335 395 320 324 340 302 367 321 393 342 376 350 313 346 392 307 317 356
#> [379] 349 385 322 386 372 311 399 368 305 381 375 357 360 353 329 374 319 370
#> [397] 379 369 306 301