Simulated 50 \(\times\) 15 Data

Data of 50 objects and 15 attributes, which were randomly generated from the 100 \(\times\) 15 simulated dataset (see data100x15). The first column contains the object names and each subsequent column is a complete ranking of the 50 objects.

data(data50x15)

Format

A data frame with 50 rows and 16 columns:

Object

object name

Ranking 1

ranking on the first attribute

Ranking 2

ranking on the second attribute

Ranking 3

ranking on the third attribute

Ranking 4

ranking on the fourth attribute

Ranking 5

ranking on the fifth attribute

Ranking 6

ranking on the sixth attribute

Ranking 7

ranking on the seventh attribute

Ranking 8

ranking on the eigth attribute

Ranking 9

ranking on the ninth attribute

Ranking 10

ranking on the tenth attribute

Ranking 11

ranking on the eleventh attribute

Ranking 12

ranking on the twelfth attribute

Ranking 13

ranking on the thirteenth attribute

Ranking 14

ranking on the fourteenth attribute

Ranking 15

ranking on the fifteenth attribute

Source

Badal, P. S., & Das, A. (2018). Efficient algorithms using subiterative convergence for Kemeny ranking problem. Computers & Operations Research, 98, 198-210. doi:10.1016/j.cor.2018.06.007

Examples

data(data50x15)
input_rkgs <- t(as.matrix(data50x15[, -1]))
obj_names <- data50x15[,1]

# Determine the mean seed ranking
mean_seed(input_rkgs)
#>  [1]  7 39 47  8 13 43 32 33 22 26 34 15 10 35 11 19 30  1 12 44 37  2 50  9 25
#> [26] 40  3 42  4 46 24 31 38 27  5 21 41 16 36  6 45 29 49 20 14 23 48 18 17 28