FUR

FUR is a heuristic algorithm to obtain a consensus ranking. It contains three branches -- Fixed, Update, and Range -- that use Subiterative Convergence and Greedy Algorithm iteratively. See `Details' for more information on each branch.

fur(input_rkgs, subit_len_list, search_radius, seed_rkg = c())

Arguments

input_rkgs

a n by k matrix of k rankings of n objects, where each column is a complete ranking.

subit_len_list

a vector containing positive integer(s) for the subiteration lengths to Subiterative Convergence. Recommended values are between 2 and 8. Smaller subiteration lengths result in shorter run-time.

search_radius

a positive integer for the maximum change in the rank of each object in the Greedy Algorithm. The default value of 0 considers all possible rank changes for each object. It is recommended to use a search radius of less than or equal to \(\min(30, \lfloor \mbox{n}/2 \rfloor)\).

seed_rkg

a vector of length n with an initial ranking to begin FUR. If the default value of an empty vector is used, then the mean seed ranking is adopted as the initial ranking to FUR.

Value

A list containing the consensus ranking (expressed as ordering), total Kemeny distance, and average tau correlation coefficient corresponding to the consensus ranking.

Details

The Fixed branch applies Subiterative Convergence using one subiteration length from subit_len_list at a time.

The Update branch executes Subiterative Convergence using the first subiteration length in subit_len_list, and then uses its output in the next call to Subiterative Convergence with the next subiteration length in the list. This process repeats until subit_len_list is exhausted.

The Range branch calls Subiterative Convergence on all subiteration lengths in subit_len_list and only retains the best ranking among these separate calls.

The output from the Subiterative Convergence calls are fed into the Greedy Algorithm as its seed ranking, and the FUR algorithm is terminated when the input to the Greedy Algorithm converges to the output and all branches have been executed at least once.

References

Badal, P. S., & Das, A. (2018). Efficient algorithms using subiterative convergence for Kemeny ranking problem. Computers & Operations Research, 98, 198-210. doi:10.1016/j.cor.2018.06.007

See also

Examples

## One subiteration length
input_rkgs <- matrix(c(3, 2, 5, 4, 1, 2, 3, 1, 5, 4, 5, 1, 3, 4, 2, 1, 2, 4, 5, 3),
    byrow = FALSE, ncol = 4)
subit_len_list <- 2
search_radius <- 1
fur(input_rkgs, subit_len_list, search_radius) # Determined the consensus ranking, total Kemeny
#> $ConsensusRanking
#> [1] 3 1 4 5 2
#> 
#> $KemenyDistance
#> [1] 22
#> 
#> $tau
#> [1] 0.45
#> 
                                              # distance, and average tau correlation coefficient

## Multiple subiteration lengths
input_rkgs <- matrix(c(3, 2, 5, 4, 1, 2, 3, 1, 5, 4, 5, 1, 3, 4, 2, 1, 2, 4, 5, 3),
    byrow = FALSE, ncol = 4)
subit_len_list <- c(2,3)
search_radius <- 1
fur(input_rkgs, subit_len_list, search_radius)
#> $ConsensusRanking
#> [1] 3 1 4 5 2
#> 
#> $KemenyDistance
#> [1] 22
#> 
#> $tau
#> [1] 0.45
#> 

## Included dataset of 15 input rankings of 50 objects
data(data50x15)
input_rkgs <- as.matrix(data50x15[, -1])
subit_len_list <- c(2, 3)
search_radius <- 1
fur(input_rkgs, subit_len_list, search_radius)
#> $ConsensusRanking
#>  [1]  4 39 48  8 12 43 32 37 24 22 34 15  9 35 10 18 30  1 13 44 40  2 50 11 21
#> [26] 41  3 36  5 49 27 31 33 28  6 26 42 16 38  7 47 29 45 25 14 20 46 17 19 23
#> 
#> $KemenyDistance
#> [1] 15222
#> 
#> $tau
#> [1] 0.1715918
#>