Mean Seed Ranking

Determine the mean seed ranking of the given input rankings. The average rank of an object is the sum of its various rankings from each input ranking divided by the total number of rankings. The mean seed ranking is formed by ranking the objects based on their average ranks, and ties are broken by ranking the first tied object with a higher rank.

mean_seed(input_rkgs)

Arguments

input_rkgs

a n by k matrix of k rankings of n objects, where each column is a complete ranking.

Value

A vector containing the mean seed ranking of the input rankings.

See also

Examples

## Four input rankings of five objects
input_rkgs <- matrix(c(3, 2, 5, 1, 2, 3, 1, 2, 5, 1, 3, 4, 4, 5, 4, 5, 1, 4, 2, 3), ncol = 5)
mean_seed(input_rkgs) # Found the mean seed ranking
#> [1] 3 1 4 5 2

## Included dataset of 15 input rankings of 50 objects
data(data50x15)
input_rkgs <- t(as.matrix(data50x15[, -1]))
mean_seed(input_rkgs)
#>  [1]  7 39 47  8 13 43 32 33 22 26 34 15 10 35 11 19 30  1 12 44 37  2 50  9 25
#> [26] 40  3 42  4 46 24 31 38 27  5 21 41 16 36  6 45 29 49 20 14 23 48 18 17 28