Greedy Algorithm for Rank Aggregation

Greedy Algorithm is a heuristic method that hunts for improved rankings by moving one object at a time (up or down). In case an object’s movement results in an improved ranking, the next object is moved with respect to this improved ranking. The process is repeated until all objects are considered once.

rap_greedy_alg(seed_rkg, input_rkgs, search_radius = 0)

Arguments

seed_rkg

an initial ranking to begin the algorithm. The algorithm is often used in conjunction with Subiterative Convergence.

input_rkgs

a n by k matrix of k rankings of n objects, where each column is a complete ranking.

search_radius

a positive integer for the maximum change in the rank of each object. The default value of 0 considers all possible rank changes for each object. Recommended value of search radius is less than or equal to \(\min(30, \lfloor \mbox{n}/2 \rfloor)\).

Value

A list containing the consensus ranking (expressed as ordering), total Kemeny distance, and average tau correlation coefficient corresponding to the consensus ranking.

References

Badal, P. S., & Das, A. (2018). Efficient algorithms using subiterative convergence for Kemeny ranking problem. Computers & Operations Research, 98, 198-210. doi:10.1016/j.cor.2018.06.007

See also

Examples

## Four input rankings of five objects
input_rkgs <- matrix(c(3, 2, 5, 4, 1, 2, 3, 1, 5, 4, 5, 1, 3, 4, 2, 1, 2, 4, 5, 3),
    byrow = FALSE, ncol = 4)
mean_seed_rkg <- mean_seed(t(input_rkgs))
rap_greedy_alg(mean_seed_rkg, input_rkgs, search_radius = 0) # Determined the consensus ranking,
#> $ConsensusRanking
#> [1] 3 1 4 5 2
#> 
#> $KemenyDistance
#> [1] 22
#> 
#> $tau
#> [1] 0.45
#> 
                                                             # total Kemeny distance, and average
                                                             # tau correlation coefficient

## Included dataset of 15 input rankings of 50 objects
data(data50x15)
input_rkgs <- as.matrix(data50x15[, -1])
mean_seed_rkg <- mean_seed(t(input_rkgs)) # Use the mean seed ranking as the seed ranking
rap_greedy_alg(mean_seed_rkg, input_rkgs, search_radius = 1)
#> $ConsensusRanking
#>  [1]  6 38 49  8 12 43 32 33 21 25 34 15 10 35 11 18 30  1 13 44 39  2 50  9 24
#> [26] 40  3 41  4 45 26 31 37 28  5 22 42 16 36  7 46 29 48 20 14 23 47 17 19 27
#> 
#> $KemenyDistance
#> [1] 15306
#> 
#> $tau
#> [1] 0.1670204
#>