SIgFUR

SIgFUR applies Seed-Based Iteration, Greedy Algorithm, and FUR in sequence for each element of subit_len_list_sbi. The mean seed ranking is used as the input to Seed-Based Iteration. The best of all output rankings from FUR is considered as the consensus ranking.

sigfur(
  input_rkgs,
  subit_len_list_sbi,
  omega_sbi,
  subit_len_list_fur,
  search_radius
)

Arguments

input_rkgs

a n by k matrix of k rankings of n objects, where each column is a complete ranking.

subit_len_list_sbi

a vector containing positive integer(s) for the subiteration lengths to Seed-Based Iteration. Recommended values are between 2 and 8. Smaller subiteration lengths result in shorter run-time.

omega_sbi

a positive integer for the number of repetitions of perturbing the seed ranking in Seed-Based Iteration. An omega_sbi value of 1 corresponds to a single application of Subiterative Convergence.

subit_len_list_fur

a vector containing positive integer(s) for the subiteration lengths to FUR.

search_radius

a positive integer for the maximum change in the rank of each object in the Greedy Algorithm and FUR. The default value of 0 considers all possible rank changes for each object. It is recommended to use a search radius of less than or equal to \(\min(30, \lfloor \mbox{n}/2 \rfloor)\).

Value

A list containing the consensus ranking (expressed as ordering), total Kemeny distance, and average tau correlation coefficient corresponding to the consensus ranking.

References

Badal, P. S., & Das, A. (2018). Efficient algorithms using subiterative convergence for Kemeny ranking problem. Computers & Operations Research, 98, 198-210. doi:10.1016/j.cor.2018.06.007

See also

Examples

## Four input rankings of five objects
input_rkgs <- matrix(c(3, 2, 5, 4, 1, 2, 3, 1, 5, 4, 5, 1, 3, 4, 2, 1, 2, 4, 5, 3),
    byrow = FALSE, ncol = 4)
subit_len_list_sbi <- c(2:3)
omega_sbi <- 10
subit_len_list_fur <- c(2:3)
search_radius <- 1
sigfur(input_rkgs, subit_len_list_sbi, omega_sbi, subit_len_list_fur, search_radius)
#> $ConsensusRanking
#> [1] 3 1 4 5 2
#> 
#> $KemenyDistance
#> [1] 22
#> 
#> $tau
#> [1] 0.45
#> 
# Determined the consensus ranking, total Kemeny distance, and average tau correlation coefficient

## Included dataset of 15 input rankings of 50 objects
data(data50x15)
input_rkgs <- as.matrix(data50x15[, -1])
subit_len_list_sbi <- c(3)
omega_sbi <- 5
subit_len_list_fur <- c(2:3)
search_radius <- 1
sigfur(input_rkgs, subit_len_list_sbi, omega_sbi, subit_len_list_fur, search_radius)
#> $ConsensusRanking
#>  [1]  6 32 48  5  9 43 29 35 24 22 36 16 11 30 12 20 38  1 14 37 40  3 50  7 19
#> [26] 41  4 39 10 49 27 33 34 28  2 26 42 17 44  8 47 31 45 25 15 21 46 18 13 23
#> 
#> $KemenyDistance
#> [1] 15220
#> 
#> $tau
#> [1] 0.1717007
#>