Subiterative Convergence

Subiterative Convergence finds the consensus ranking by iteratively applying the Modified Kemeny algorithm on smaller number of objects, \(\eta\). Starting with a given seed ranking, the consensus ranking is obtained when the algorithm converges.

subit_convergence(eta, seed_rkg, input_rkgs, universe_rkgs = c())

Arguments

eta

a subiteration length of number of objects to consider in the smaller subset. Recommended eta values are between 2 and 8. Smaller eta values result in shorter run-time.

seed_rkg

an initial ranking to start the algorithm. An ideal seed ranking for Subiterative Convergence is the mean seed ranking of input rankings.

input_rkgs

a n by k matrix of k rankings of n objects, where each column is a complete ranking.

universe_rkgs

a matrix containing all possible permutations of ranking n objects. Each column in this matrix represents one permuted ranking.

Value

A list containing the consensus ranking (expressed as ordering), total Kemeny distance, and average tau correlation coefficient corresponding to the consensus ranking.

References

Badal, P. S., & Das, A. (2018). Efficient algorithms using subiterative convergence for Kemeny ranking problem. Computers & Operations Research, 98, 198-210. doi:10.1016/j.cor.2018.06.007

See also

Examples

## Four input rankings of five objects
eta <- 3
seed_rkg <- c(1, 2, 3, 4, 5)
input_rkgs <- matrix(c(3, 2, 5, 4, 1, 2, 3, 1, 5, 4, 5, 1, 3, 4, 2, 1, 2, 4, 5, 3),
    byrow = FALSE, ncol = 4)
subit_convergence(eta, seed_rkg, input_rkgs) # Determined the consensus ranking, total Kemeny
#> $ConsensusRanking
#> [1] 2 1 4 5 3
#> 
#> $KemenyDistance
#> [1] 22
#> 
#> $tau
#> [1] 0.45
#> 
                                             # distance, and average tau correlation coefficient

## Example with eta=1
eta <- 1
seed_rkg <- c(1, 2, 3, 4, 5)
input_rkgs <- matrix(c(3, 2, 5, 4, 1, 2, 3, 1, 5, 4, 5, 1, 3, 4, 2, 1, 2, 4, 5, 3),
    byrow = FALSE, ncol = 4)
subit_convergence(eta, seed_rkg, input_rkgs) # Shows a warning and returns seed ranking
#> Warning: Caution: eta=1 has no effect on subiteration converence, so the intial starting ranking is returned.
#> $ConsensusRanking
#> [1] 1 2 3 4 5
#> 
#> $KemenyDistance
#> [1] 34
#> 
#> $tau
#> [1] 0.15
#> 

## Included dataset of 15 input rankings of 50 objects
data(data50x15)
input_rkgs <- as.matrix(data50x15[, -1])
mean_seed_rkg <- mean_seed(t(input_rkgs)) # Use the mean seed ranking as the seed ranking
eta <- 2
subit_convergence(eta, seed_rkg = mean_seed_rkg, input_rkgs)
#> $ConsensusRanking
#>  [1]  6 39 49  8 12 43 32 34 20 25 35 15 10 36 11 18 30  1 13 45 40  2 50  9 24
#> [26] 41  3 37  4 47 26 31 33 28  5 22 42 16 38  7 48 29 46 21 14 23 44 17 19 27
#> 
#> $KemenyDistance
#> [1] 15262
#> 
#> $tau
#> [1] 0.169415
#>